Fixed breakpoints loop conditions
The validate breakpoints loop condition was not working because it used in instead of of in the loop declaration which meant itlooped the keys of the Array object and not the elements of the object which always signaled an error for the first item in array due to being a 0 index key. The loop to check if a breakpoint matched the current window width was not working as expected because the condition evaluated anything bigger and not equal which meant that when the window width was exactly the same as the breakpoint it was not being considered as the breakpoint.