Add a custom hook that allows control of HTML input element pickers
By returning two hook callbacks to the calling component it can controll when to open the picker and give accesibility to open the picker with the keyboard. This is specially important since the inputs have additional elements beside just the html element that should also open the picker if clicked.